% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterGo.R
\name{ClusterGo}
\alias{ClusterGo}
\title{Perform Gene Ontology (GO) Enrichment Analysis for Clusters}
\usage{
ClusterGo(
  Deg_result,
  org_db = "org.Hs.eg.db",
  ont = "BP",
  p_val_cutoff = 0.05,
  q_val_cutoff = 0.02
)
}
\arguments{
\item{Deg_result}{A list object typically returned by `FindClusterDeg`.
It must contain a sub-list named `by_cluster`, where each element contains
a `gene` slot.}

\item{org_db}{The organism database object or name.}

\item{ont}{One of "BP", "MF", or "CC".}

\item{p_val_cutoff}{P-value cutoff for enrichment.}

\item{q_val_cutoff}{Q-value cutoff for enrichment.}
}
\value{
A named list of enrichment results (`enrichResult` objects).
Clusters with no significant enrichment are stored as NA.
}
\description{
Iterates through a list of cluster markers (typically the output of
FindClusterDeg) and performs GO enrichment analysis using clusterProfiler.
}
\examples{
# Example 1 using subset of pbmc data
# Complete pre-processing prior clustering
sub_pbmc <- Seurat::ScaleData(sub_pbmc)
sub_pbmc <- Seurat::RunPCA(sub_pbmc)
sub_pbmc <- Seurat::FindNeighbors(sub_pbmc, dims = 1:15, reduction = "pca")
# perform clustering using default parameter without plotting
sub_pbmc <- ClusterUnderRes(sub_pbmc, showPlot = FALSE)
# Set clustering result under resolution = 0.5
sub_pbmc <- Seurat::FindClusters(sub_pbmc, resolution = 0.5)
# Find Cluster DEGs
DegResult <- FindClusterDeg(sub_pbmc)
# Perform GO enrichment analysis on Cluster DEGs
GoResult <- ClusterGo(DegResult)

\dontrun{
# Example 2 using pbmc data
# Complete pre-processing prior clustering
sub_pbmc <- Seurat::ScaleData(sub_pbmc)
sub_pbmc <- Seurat::RunPCA(sub_pbmc)
sub_pbmc <- Seurat::FindNeighbors(sub_pbmc, dims = 1:15, reduction = "pca")
# perform clustering using default parameter without plotting
sub_pbmc <- ClusterUnderRes(sub_pbmc, showPlot = FALSE)
# Set clustering result under resolution = 0.5
sub_pbmc <- Seurat::FindClusters(sub_pbmc, resolution = 0.5)
# Find Cluster DEGs
DegResult <- FindClusterDeg(sub_pbmc)
# Perform GO enrichment analysis on Cluster DEGs
GoResult <- ClusterGo(DegResult)
}

}
\references{
Wu, T., Hu, E., Xu, S., Chen, M., Guo, P., Dai, Z., Feng, T., Zhou, L., Tang, W., Zhan, L., Fu, X., Liu, S., Bo, X., & Yu, G. (2021). clusterProfiler 4.0: A universal enrichment tool for interpreting omics data. The Innovation, 2(3). \href{https://doi.org/10.1016/j.xinn.2021.100141}{Link}
}
