% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterUnderRes.R
\name{ClusterUnderRes}
\alias{ClusterUnderRes}
\title{Cluster Seurat object under given range of resolution}
\usage{
ClusterUnderRes(obj, start = 0, end = 1.5, margin = 0.1, showPlot = TRUE)
}
\arguments{
\item{obj}{a Seurat object after pro-processing, PCA, and Neighbors constructed}

\item{start}{The start of range of resolution (>= 0)}

\item{end}{The end of range of resolution, inclusive (>= start)}

\item{margin}{The margin of range (positive rational number)}

\item{showPlot}{whether display the clustering tree plot (T/F)}
}
\value{
Returns an Seurat object with clustering result stored in "metadata"
slot with naming convention "RNA_snn_res.<resolution>".
}
\description{
Utilize Clustree, plot clustering tree showing the relationship
between clustering under various resolution and return clustered Seurat object
with clustering result stored in "metadata" slot with naming convention
"RNA_snn_res.<resolution>".
}
\examples{
# example using pbmc data
# Complete pre-processing prior clustering
pbmc <- Seurat::ScaleData(pbmc)
pbmc <- Seurat::RunPCA(pbmc)
pbmc <- Seurat::FindNeighbors(pbmc, dims = 1:15, reduction = "pca")
# perform clustering using default parameter without plotting
pbmc <- ClusterUnderRes(pbmc, showPlot = FALSE)

}
\references{
Zappia, L., & Oshlack, A. (2018). Clustering trees: A visualization for evaluating clusterings at multiple resolutions. GigaScience, 7(7), giy083. \href{https://doi.org/10.1093/gigascience/giy083}{Link}

Hao, Y., Stuart, T., Kowalski, M. H., Choudhary, S., Hoffman, P., Hartman, A., Srivastava, A., Molla, G., Madad, S., Fernandez-Granda, C., & Satija, R. (2024). Dictionary learning for integrative, multimodal and scalable single-cell analysis. Nature Biotechnology, 42(2), 293–304. \href{https://doi.org/10.1038/s41587-023-01767-y}{Link}
}
