% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustifyrAnnote.R
\name{ClustifyrAnnote}
\alias{ClustifyrAnnote}
\title{Annotate cell using Clustifyr package}
\usage{
ClustifyrAnnote(obj, ref_mat, cluster_col = "seurat_clusters")
}
\arguments{
\item{obj}{a Seurat object after pre-processing, PCA, and Neighbors constructed}

\item{ref_mat}{A reference expression matrix (https://www.bioconductor.org/packages/release/data/experiment/vignettes/clustifyrdatahub/inst/doc/clustifyrdatahub.html).}

\item{cluster_col}{The name of the metadata column containing the cluster identities to annotate. Defaults to "seurat_clusters".}
}
\value{
Returns an Seurat object with Clustifyr annotation result added into
metadata slot as Clustifyr.labels
}
\description{
Utilize Clustifyr, annotate cell automatically at cluster level by map to
selected previously annotated scRNA-seq data. Annotation result will be stored
in metadata slot as Clustifyr.labels
}
\examples{

# Example 1 using subset of pbmc data
# Complete pre-processing prior clustering
sub_pbmc <- Seurat::ScaleData(sub_pbmc)
sub_pbmc <- Seurat::RunPCA(sub_pbmc)
sub_pbmc <- Seurat::FindNeighbors(sub_pbmc, dims = 1:15, reduction = "pca")
sub_pbmc <- Seurat::FindClusters(sub_pbmc, resolution = 0.5)
sub_pbmc <- Seurat::RunUMAP(sub_pbmc, dims = 1:15)
# perform annotation using Human hematopoietic cell microarray
# (https://ftp.ncbi.nlm.nih.gov/geo/series/GSE24nnn/GSE24759/matrix/GSE24759_series_matrix.txt.gz)
sub_pbmc <- ClustifyrAnnote(sub_pbmc, clustifyrdatahub::ref_hema_microarray())

\dontrun{
# Example 2 using pbmc data
# Complete pre-processing prior clustering
pbmc <- Seurat::ScaleData(pbmc)
pbmc <- Seurat::RunPCA(pbmc)
pbmc <- Seurat::FindNeighbors(pbmc, dims = 1:15, reduction = "pca")
pbmc <- Seurat::FindClusters(pbmc, resolution = 0.5)
pbmc <- Seurat::RunUMAP(pbmc, dims = 1:15)
# perform annotation using Human hematopoietic cell microarray
# (https://ftp.ncbi.nlm.nih.gov/geo/series/GSE24nnn/GSE24759/matrix/GSE24759_series_matrix.txt.gz)
pbmc <- ClustifyrAnnote(pbmc, clustifyrdatahub::ref_hema_microarray())
}

}
\references{
Hao, Y., Stuart, T., Kowalski, M. H., Choudhary, S., Hoffman, P., Hartman, A., Srivastava, A., Molla, G., Madad, S., Fernandez-Granda, C., & Satija, R. (2024). Dictionary learning for integrative, multimodal and scalable single-cell analysis. Nature Biotechnology, 42(2), 293–304. \href{https://doi.org/10.1038/s41587-023-01767-y}{Link}

Fu, R., Gillen, A. E., Sheridan, R. M., Tian, C., Daya, M., Hao, Y., Hesselberth, J. R., & Riemondy, K. A. (2020). clustifyr: An R package for automated single-cell RNA sequencing cluster classification. F1000Research. \href{https://doi.org/10.12688/f1000research.22969.2}{Link}
}
