% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClusterDeg.R
\name{FindClusterDeg}
\alias{FindClusterDeg}
\title{Find, filter, and score cluster markers (Differential experssed genes)}
\usage{
FindClusterDeg(
  obj,
  only_pos = TRUE,
  pct_min = 0.1,
  pval_max = 0.05,
  logfc_min = 0.25
)
}
\arguments{
\item{obj}{Seurat object with identities set (Idents(obj)).}

\item{only_pos}{logical; pass to FindAllMarkers (default TRUE).}

\item{pct_min}{Minimum fraction expressed in each group (applied to both pct.1 and pct.2).}

\item{pval_max}{Maximum raw p-value to keep.}

\item{logfc_min}{Keep rows with |avg_log2FC| >= logfc_min.}
}
\value{
A list with:
\describe{
  \item{by_cluster}{named list of filtered marker data.frames per cluster}
  \item{combined}{single data.frame of filtered markers with p_FC}
}
}
\description{
Runs Seurat::FindAllMarkers(), filters markers by prevalence, p-value,
and log2FC thresholds, then computes a p_FC score = (1 - p_val) * avg_log2FC.
Return a list of length two compose of splitted and combined DEG result
}
\examples{
# example using pbmc data
# Complete pre-processing prior clustering
pbmc <- Seurat::ScaleData(pbmc)
pbmc <- Seurat::RunPCA(pbmc)
pbmc <- Seurat::FindNeighbors(pbmc, dims = 1:15, reduction = "pca")
# perform clustering using default parameter without plotting
pbmc <- ClusterUnderRes(pbmc, showPlot = FALSE)
# Set clustering result under resolution = 0.5
pbmc <- Seurat::FindClusters(pbmc, resolution = 0.5)
# Find Cluster DEGs
result <- FindClusterDeg(pbmc)

}
\references{
Hao, Y., Stuart, T., Kowalski, M. H., Choudhary, S., Hoffman, P., Hartman, A., Srivastava, A., Molla, G., Madad, S., Fernandez-Granda, C., & Satija, R. (2024). Dictionary learning for integrative, multimodal and scalable single-cell analysis. Nature Biotechnology, 42(2), 293–304. \href{https://doi.org/10.1038/s41587-023-01767-y}{Link}
}
