% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{10X Genomics Peripheral Blood Mononuclear Cells' (PBMC) scRNA-seq data}
\format{
A Seurat object with the following slots filled:
\describe{
 \item{assays}{RNA expression data}
 \item{meta.data}{Cell level metadata}
 \item{active.ident}{current cell identify}
 \item{version}{Seurat verison used to create this obejct}
 \item{commands}{Command history}
}
}
\source{
10X Genomics, Inc
}
\usage{
pbmc
}
\description{
Consist of 2638 cell's metadata and RNA expression data
post quality control and normalization from Seurat's vignettes
(https://satijalab.org/seurat/articles/pbmc3k_tutorial)
}
\examples{
\dontrun{
 pbmc
}
}
\keyword{datasets}
