% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscClustAnnot.R
\name{runscClustAnnot}
\alias{runscClustAnnot}
\title{Launch Shiny App for scClustAnnot}
\usage{
runscClustAnnot()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for scClustAnnot.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

scClustAnnot::runscClustAnnot()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
