% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{sub_pbmc}
\alias{sub_pbmc}
\title{Random Subset of 10X Genomics Peripheral Blood Mononuclear Cells (PBMC) scRNA-seq data}
\format{
A Seurat object with the following slots filled:
\describe{
  \item{assays}{RNA expression data (contains 'counts' and 'data' layers)}
  \item{meta.data}{Basic cell-level metadata (e.g., nFeature_RNA, nCount_RNA)}
  \item{active.ident}{Default identity (typically set to project name, as no clusters are defined)}
  \item{version}{Seurat version used to create this object}
  \item{commands}{Command history (subsetting and pre-processing steps)}
}
}
\source{
10X Genomics, Inc
}
\usage{
sub_pbmc
}
\description{
Consist of 500 cell's metadata and RNA expression data
post quality control from Seurat's vignettes
(https://satijalab.org/seurat/articles/pbmc3k_tutorial)
}
\examples{
\dontrun{
 sub_pbmc
}
}
\keyword{datasets}
